refactor: 💡 type and constanse move to type file
