+++ src/pages/contact/logic/judgmentErrorState.ts
@@ -1,5 +1,4 @@
-import { ErrorState } from '@/pages/contact/index.page';
-import { InitialInput } from '@/pages/contact/type';
+import { ErrorState, InitialInput } from '@/pages/contact/type';
 
 export const judgmentErrorState = (input: InitialInput) => {
   const newErrors: ErrorState = {};
