+++ src/pages/contact/type/index.ts
@@ -7,3 +7,15 @@ export type InitialInput = {
   material: string;
   postContent: string;
 };
+
+export const initialInput = {
+  name: '',
+  katakana: '',
+  mail: '',
+  phone: '',
+  period: '1',
+  material: '',
+  postContent: '',
+};
+
+export type ErrorState = Partial<Record<keyof InitialInput, string>>;
